image sharpening/requirements.txt
torch
torchvision
opencv-python
matplotlib
tqdm
Pillow